Changed package name and class names for better structure
