--- employ-easy/src/test/java/se/yrgo/employee/entities/enums/EmployeeStatusTest.java
@@ -1,22 +1,0 @@
-package se.yrgo.employee.entities.enums;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.junit.jupiter.api.Test;
-
-class EmployeeStatusTest {
-
-    @Test
-    void valueOfThrowsIllegalArg() {
-        IllegalArgumentException thrown = assertThrows(
-            IllegalArgumentException.class,
-            () -> EmployeeStatus.valueOf(-1)
-        );
-        assertEquals("Invalid Employee Status code.", thrown.getMessage());
-    }
-
-    @Test
-    void valueOfReturnsCorrectly() {
-        assertEquals(EmployeeStatus.ACTIVE, EmployeeStatus.valueOf(1));
-    }
-}
